fix(dogfood): closed captions appeared only on speaking while muted
