+++ sample-apps/react-native/dogfood/src/components/CallControls/BottomControls/BadgeCountIndicator.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { useTheme } from '@stream-io/video-react-native-sdk';
-import { Z_INDEX } from '../../constants';
-import { ComponentTestIds } from '../../constants/TestIds';
+import { Z_INDEX } from '../../../constants';
+import { ComponentTestIds } from '../../../constants/TestIds';
 
 /**
  * A badge that displays a number.
