+++ sample-apps/react-native/dogfood/src/components/CallControls/BottomControls/ParticipantsButton.tsx
@@ -7,7 +7,7 @@ import {
 } from '@stream-io/video-react-native-sdk';
 import { IconWrapper } from '@stream-io/video-react-native-sdk/src/icons';
 import { BadgeCountIndicator } from './BadgeCountIndicator';
-import Participants from '../../assets/Participants';
+import Participants from '../../../assets/Participants';
 
 /**
  * The props for the Participants Button in the Call Controls.
