+++ sample-apps/react-native/dogfood/src/components/CallControls/TopControls/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useMemo } from 'react';
 import { View, StyleSheet } from 'react-native';
-import { TopViewBackground } from '../../assets/TopViewBackground';
+import { TopViewBackground } from '../../../assets/TopViewBackground';
 import {
   HangUpCallButton,
   useTheme,
